pc_test_python/requirements.txt
opencv-python
numpy
pyyaml
ultralytics
streamlit
pandas
